backend/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
chardet==6.0.0.post1
google-genai==0.1.0
flask-cors==4.0.0
supabase==2.4.0
postgrest-py==0.10.6
realtime-py==0.1.3
python-dotenv==1.0.0
PyJWT==2.12.1
bcrypt==4.1.2